NIFI-10953: add more description details and default value for json payload
