+++ nifi-nar-bundles/nifi-gcp-bundle/nifi-gcp-processors/src/main/java/org/apache/nifi/processors/gcp/vision/AbstractStartGcpVisionOperation.java
@@ -17,8 +17,6 @@
 
 package org.apache.nifi.processors.gcp.vision;
 
-import static org.apache.nifi.processors.gcp.util.GoogleUtils.GCP_CREDENTIALS_PROVIDER_SERVICE;
-
 import com.google.api.gax.longrunning.OperationFuture;
 import com.google.protobuf.util.JsonFormat;
 import java.io.ByteArrayInputStream;
@@ -26,39 +24,19 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 import org.apache.nifi.annotation.lifecycle.OnStopped;
 import org.apache.nifi.components.PropertyDescriptor;
-import org.apache.nifi.expression.ExpressionLanguageScope;
 import org.apache.nifi.flowfile.FlowFile;
 import org.apache.nifi.processor.ProcessContext;
 import org.apache.nifi.processor.ProcessSession;
 import org.apache.nifi.processor.exception.ProcessException;
-import org.apache.nifi.processor.util.StandardValidators;
 
 public abstract class AbstractStartGcpVisionOperation<B extends com.google.protobuf.GeneratedMessageV3.Builder<B>> extends AbstractGcpVisionProcessor  {
-    public static final PropertyDescriptor JSON_PAYLOAD = new PropertyDescriptor.Builder()
-            .name("json-payload")
-            .displayName("JSON Payload")
-            .description("JSON request for AWS Machine Learning services. The Processor will use FlowFile content for the request when this property is not specified.")
-            .expressionLanguageSupported(ExpressionLanguageScope.FLOWFILE_ATTRIBUTES)
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .build();
-    private static final List<PropertyDescriptor> PROPERTIES = Collections.unmodifiableList(Arrays.asList(
-            JSON_PAYLOAD, GCP_CREDENTIALS_PROVIDER_SERVICE));
-
-    @Override
-    public List<PropertyDescriptor> getSupportedPropertyDescriptors() {
-        return PROPERTIES;
-    }
 
     @Override
     public void onTrigger(ProcessContext context, ProcessSession session) throws ProcessException {
         FlowFile flowFile = session.get();
-        if (flowFile == null && !context.getProperty(JSON_PAYLOAD).isSet()) {
+        if (flowFile == null && !context.getProperty(getJsonPayloadPropertyDescriptor()).isSet()) {
             return;
         } else if (flowFile == null) {
             flowFile = session.create();
@@ -81,7 +59,8 @@ public abstract class AbstractStartGcpVisionOperation<B extends com.google.proto
 
     protected OperationFuture<?, ?> startOperation(ProcessSession session, ProcessContext context, FlowFile flowFile) {
         B builder = newBuilder();
-        InputStream inStream = context.getProperty(JSON_PAYLOAD).isSet() ? getInputStreamFromProperty(context) : session.read(flowFile);
+        InputStream inStream = context.getProperty(getJsonPayloadPropertyDescriptor()).isSet()
+                ? getInputStreamFromProperty(context, flowFile) : session.read(flowFile);
         try (InputStream inputStream = inStream) {
             JsonFormat.parser().ignoringUnknownFields().merge(new InputStreamReader(inputStream), builder);
         } catch (final IOException e) {
@@ -90,11 +69,13 @@ public abstract class AbstractStartGcpVisionOperation<B extends com.google.proto
         return startOperation(builder);
     }
 
-    private InputStream getInputStreamFromProperty(ProcessContext context) {
-        return new ByteArrayInputStream(context.getProperty(JSON_PAYLOAD).getValue().getBytes(StandardCharsets.UTF_8));
+    private InputStream getInputStreamFromProperty(ProcessContext context, FlowFile flowFile) {
+        return new ByteArrayInputStream(context.getProperty(getJsonPayloadPropertyDescriptor()).evaluateAttributeExpressions(flowFile).getValue().getBytes(StandardCharsets.UTF_8));
     }
 
     abstract B newBuilder();
 
     abstract OperationFuture<?, ?> startOperation(B builder);
+
+    abstract PropertyDescriptor getJsonPayloadPropertyDescriptor();
 }
