+++ nifi-nar-bundles/nifi-gcp-bundle/nifi-gcp-processors/src/main/java/org/apache/nifi/processors/gcp/vision/StartGcpVisionAnnotateImagesOperation.java
@@ -17,13 +17,21 @@
 
 package org.apache.nifi.processors.gcp.vision;
 
+import static org.apache.nifi.processors.gcp.util.GoogleUtils.GCP_CREDENTIALS_PROVIDER_SERVICE;
+
 import com.google.api.gax.longrunning.OperationFuture;
 import com.google.cloud.vision.v1.AsyncBatchAnnotateImagesRequest;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 import org.apache.nifi.annotation.behavior.WritesAttribute;
 import org.apache.nifi.annotation.behavior.WritesAttributes;
 import org.apache.nifi.annotation.documentation.CapabilityDescription;
 import org.apache.nifi.annotation.documentation.SeeAlso;
 import org.apache.nifi.annotation.documentation.Tags;
+import org.apache.nifi.components.PropertyDescriptor;
+import org.apache.nifi.expression.ExpressionLanguageScope;
+import org.apache.nifi.processor.util.StandardValidators;
 
 @Tags({"Google", "Cloud", "Machine Learning", "Vision"})
 @CapabilityDescription("Trigger a Vision operation on image input. It should be followed by GetGcpVisionAnnotateImagesOperationStatus processor in order to monitor operation status.")
@@ -32,6 +40,40 @@ import org.apache.nifi.annotation.documentation.Tags;
         @WritesAttribute(attribute = "operationKey", description = "Unique key of the operation.")
 })
 public class StartGcpVisionAnnotateImagesOperation extends AbstractStartGcpVisionOperation<AsyncBatchAnnotateImagesRequest.Builder> {
+    static final PropertyDescriptor JSON_PAYLOAD = new PropertyDescriptor.Builder()
+            .name("json-payload")
+            .displayName("JSON Payload")
+            .description("JSON request for AWS Machine Learning services. The Processor will use FlowFile content for the request when this property is not specified.")
+            .expressionLanguageSupported(ExpressionLanguageScope.FLOWFILE_ATTRIBUTES)
+            .required(false)
+            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
+            .defaultValue("{\n" +
+                    "    \"requests\": [{\n" +
+                    "        \"image\": {\n" +
+                    "            \"source\": {\n" +
+                    "                \"imageUri\": \"gs://#{bucket}/${filename}\"\n" +
+                    "            }\n" +
+                    "        },\n" +
+                    "        \"features\": [{\n" +
+                    "            \"type\": \"FACE_DETECTION\",\n" +
+                    "            \"maxResults\": 4\n" +
+                    "        }]\n" +
+                    "    }],\n" +
+                    "    \"outputConfig\": {\n" +
+                    "        \"gcsDestination\": {\n" +
+                    "            \"uri\": \"gs://#{bucket}/${filename}/\"\n" +
+                    "        },\n" +
+                    "        \"batchSize\": 2\n" +
+                    "    }")
+            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
+            .build();
+    private static final List<PropertyDescriptor> PROPERTIES = Collections.unmodifiableList(Arrays.asList(
+            JSON_PAYLOAD, GCP_CREDENTIALS_PROVIDER_SERVICE));
+
+    @Override
+    public List<PropertyDescriptor> getSupportedPropertyDescriptors() {
+        return PROPERTIES;
+    }
 
     @Override
     AsyncBatchAnnotateImagesRequest.Builder newBuilder() {
@@ -42,4 +84,10 @@ public class StartGcpVisionAnnotateImagesOperation extends AbstractStartGcpVisio
     OperationFuture<?, ?> startOperation(AsyncBatchAnnotateImagesRequest.Builder builder) {
         return getVisionClient().asyncBatchAnnotateImagesAsync(builder.build());
     }
+
+    @Override
+    PropertyDescriptor getJsonPayloadPropertyDescriptor() {
+        return JSON_PAYLOAD;
+    }
+
 }
