+++ nifi-nar-bundles/nifi-gcp-bundle/nifi-gcp-processors/src/main/resources/docs/org.apache.nifi.processors.gcp.vision.StartGcpVisionAnnotateFilesOperation/additionalDetails.html
@@ -49,7 +49,7 @@
         {
             "inputConfig": {
                 "gcsSource": {
-                    "uri": "gs://..."
+                    "uri": "gs://#{bucket}/${filename}"
                 },
                 "mimeType": "application/pdf"
             },
@@ -59,7 +59,7 @@
                 }],
             "outputConfig": {
                 "gcsDestination": {
-                    "uri": "gs://..."
+                    "uri": "gs://#{bucket}/${filename}/"
                 },
                 "batchSize": 2
             }
@@ -74,7 +74,32 @@
 </ul>
 You can find more details on <a href="https://cloud.google.com/vision/docs/features-list" target="_blank">Google Vision Feature List</a>
 
-<h3>Example Flow</h3>
+<h3>Example: How to setup simple Annotate Image Flow</h3>
+<p>
+    Prerequisites
+</p>
+<p>
+    <ul>
+    <li>Make sure files are available in GCS</li>
+    <li>Setup a parameter context with <code>bucket</code> parameter and its value will be the name of your GCS bucket</li>
+    <li>Bucket must contain only your files.</li>
+    <li>Set Process Group Parameter Context</li>
+    </ul>
+</p>
+<p>Create the following flow</p>
 <img src="vision-annotate-files.png" style="height: 50%; width: 50%"/>
+<p>
+Keep the default value of JSON PAYLOAD property in StartGcpVisionAnnotateImagesOperation
+</p>
+<p>
+Execution steps:
+    <ul>
+        <li>ListGCSBucket processor will return list of files in the bucket for the first trigger.</li>
+        <li>ListGCSBucket will return new items since the last trigger for the additional runs</li>
+        <li>StartGcpVisionAnnotateFilesOperation processor will trigger GCP Vision jobs based on the json configuration.</li>
+        <li>StartGcpVisionAnnotateFilesOperation processor populate <code>operationKey</code> flow file attribute</li>
+        <li>GetGcpVisionAnnotateFilesOperationStatus processor will periodically query status of the job by the <code>operationKey</code>.</li>
+    </ul>
+</p>
 </body>
 </html>
